Download Darwin Cloud Data

`download_cloud_data()` is a function used to download Darwin Cloud data. First it downloads data, then it subsets informative columns and renames them so they could be used in `shiny` app.

download_cloud_data(
  path_remote = "https://raw.githubusercontent.com/kurator-org/",
  path_github = "kurator-validation/master/packages/kurator_dwca/",
  path_file = "data/vocabularies/darwin_cloud.txt",
  column_field = "fieldname",
  column_stand = "standard"
)

Arguments

path_remote
a character string that specifies path to remote repository (Kurator in Github)

path_github
a character string that specifies path within given repository

path_file
a character string that specifies name of a dictionary file

column_field
a character string that specifies name of the column that contains field information

column_stand
a character string that specifies name of the column that contains standard information

Value

data.frame of Darwin Cloud data.

See also

Examples

download_cloud_data()

#>                                   fieldname                            standard
#> 1                         ACCEPTEDNAMEUSAGE                   acceptedNameUsage
#> 2                       ACCEPTEDNAMEUSAGEID                 acceptedNameUsageID
#> 3                              ACCESSRIGHTS                        accessRights
#> 4                                       ANO                                year
#> 5                                       AÑO                                year
#> 6                           ASSOCIATEDMEDIA                     associatedMedia
#> 7                     ASSOCIATEDOCCURRENCES               associatedOccurrences
#> 8                       ASSOCIATEDORGANISMS                 associatedOrganisms
#> 9                      ASSOCIATEDREFERENCES                associatedReferences
#> 10                      ASSOCIATEDSEQUENCES                 associatedSequences
#> 11                           ASSOCIATEDTAXA                      associatedTaxa
#> 13                            BASISOFRECORD                       basisOfRecord
#> 15                                      BED                                 bed
#> 16                                 BEHAVIOR                            behavior
#> 17                    BIBLIOGRAPHICCITATION               bibliographicCitation
#> 18                            CATALOGNUMBER                       catalogNumber
#> 19                                   CATNUM                       catalogNumber
#> 20                                    CLASE                               class
#> 21                                    CLASS                               class
#> 22                                 COLECTOR                          recordedBy
#> 23                           COLLECTIONCODE                      collectionCode
#> 24                             COLLECTIONID                        collectionID
#> 25                                COLLECTOR                          recordedBy
#> 26                                CONTINENT                           continent
#> 28                               CONTINENTE                           continent
#> 29                              COORDENADAS                 verbatimCoordinates
#> 30                      COORDINATEPRECISION                 coordinatePrecision
#> 31                              COORDINATES                 verbatimCoordinates
#> 32            COORDINATEUNCERTAINTYINMETERS       coordinateUncertaintyInMeters
#> 33                                  COUNTRY                             country
#> 34                              COUNTRYCODE                         countryCode
#> 35                                   COUNTY                              county
#> 36                                   CUENCA                           waterBody
#> 37                      DATAGENERALIZATIONS                 dataGeneralizations
#> 38                                DATASETID                           datasetID
#> 39                              DATASETNAME                         datasetName
#> 41                           DATEIDENTIFIED                      dateIdentified
#> 42                                    DATUM                       geodeticDatum
#> 43                                      DAY                                 day
#> 44                             DAYCOLLECTED                                 day
#> 45                            DAYIDENTIFIED                       dayIdentified
#> 46                     DCTERMS:ACCESSRIGHTS                        accessRights
#> 47            DCTERMS:BIBLIOGRAPHICCITATION               bibliographicCitation
#> 48                         DCTERMS:LANGUAGE                            language
#> 49                          DCTERMS:LICENSE                             license
#> 50                         DCTERMS:MODIFIED                            modified
#> 51                       DCTERMS:REFERENCES                          references
#> 52                     DCTERMS:RIGHTSHOLDER                        rightsHolder
#> 53                             DCTERMS:TYPE                                type
#> 54                          DECIMALLATITUDE                     decimalLatitude
#> 55                         DECIMALLONGITUDE                    decimalLongitude
#> 56                            DETERMINED BY                        identifiedBy
#> 57                                      DIA                                 day
#> 58                                      DÍA                                 day
#> 59                              DISPOSITION                         disposition
#> 60                                 DRAINAGE                           waterBody
#> 61                    DWC:ACCEPTEDNAMEUSAGE                   acceptedNameUsage
#> 62                  DWC:ACCEPTEDNAMEUSAGEID                 acceptedNameUsageID
#> 63                      DWC:ASSOCIATEDMEDIA                     associatedMedia
#> 64                DWC:ASSOCIATEDOCCURRENCES               associatedOccurrences
#> 65                  DWC:ASSOCIATEDORGANISMS                 associatedOrganisms
#> 66                 DWC:ASSOCIATEDREFERENCES                associatedReferences
#> 67                  DWC:ASSOCIATEDSEQUENCES                 associatedSequences
#> 68                       DWC:ASSOCIATEDTAXA                      associatedTaxa
#> 69                        DWC:BASISOFRECORD                       basisOfRecord
#> 70                                  DWC:BED                                 bed
#> 71                             DWC:BEHAVIOR                            behavior
#> 72                        DWC:CATALOGNUMBER                       catalogNumber
#> 73                                DWC:CLASS                               class
#> 74                       DWC:COLLECTIONCODE                      collectionCode
#> 75                         DWC:COLLECTIONID                        collectionID
#> 76                            DWC:CONTINENT                           continent
#> 77                  DWC:COORDINATEPRECISION                 coordinatePrecision
#> 78        DWC:COORDINATEUNCERTAINTYINMETERS       coordinateUncertaintyInMeters
#> 79                              DWC:COUNTRY                             country
#> 80                          DWC:COUNTRYCODE                         countryCode
#> 81                               DWC:COUNTY                              county
#> 82                  DWC:DATAGENERALIZATIONS                 dataGeneralizations
#> 83                            DWC:DATASETID                           datasetID
#> 84                          DWC:DATASETNAME                         datasetName
#> 85                       DWC:DATEIDENTIFIED                      dateIdentified
#> 86                                  DWC:DAY                                 day
#> 87                      DWC:DECIMALLATITUDE                     decimalLatitude
#> 88                     DWC:DECIMALLONGITUDE                    decimalLongitude
#> 89                          DWC:DISPOSITION                         disposition
#> 90                    DWC:DYNAMICPROPERTIES                   dynamicProperties
#> 91             DWC:EARLIESTAGEORLOWESTSTAGE            earliestAgeOrLowestStage
#> 92          DWC:EARLIESTEONORLOWESTEONOTHEM         earliestEonOrLowestEonothem
#> 93          DWC:EARLIESTEPOCHORLOWESTSERIES         earliestEpochOrLowestSeries
#> 94           DWC:EARLIESTERAORLOWESTERATHEM          earliestEraOrLowestErathem
#> 95         DWC:EARLIESTPERIODORLOWESTSYSTEM        earliestPeriodOrLowestSystem
#> 96                         DWC:ENDDAYOFYEAR                        endDayOfYear
#> 97                   DWC:ESTABLISHMENTMEANS                  establishmentMeans
#> 98                            DWC:EVENTDATE                           eventDate
#> 99                              DWC:EVENTID                             eventID
#> 100                        DWC:EVENTREMARKS                        eventRemarks
#> 101                           DWC:EVENTTIME                           eventTime
#> 102                              DWC:FAMILY                              family
#> 103                          DWC:FIELDNOTES                          fieldNotes
#> 104                         DWC:FIELDNUMBER                         fieldNumber
#> 105                 DWC:FOOTPRINTSPATIALFIT                 footprintSpatialFit
#> 106                        DWC:FOOTPRINTSRS                        footprintSRS
#> 107                        DWC:FOOTPRINTWKT                        footprintWKT
#> 108                           DWC:FORMATION                           formation
#> 109                               DWC:GENUS                               genus
#> 110                       DWC:GEODETICDATUM                       geodeticDatum
#> 111                 DWC:GEOLOGICALCONTEXTID                 geologicalContextID
#> 112                     DWC:GEOREFERENCEDBY                     georeferencedBy
#> 113                   DWC:GEOREFERENCEDDATE                   georeferencedDate
#> 114                DWC:GEOREFERENCEPROTOCOL                georeferenceProtocol
#> 115                 DWC:GEOREFERENCEREMARKS                 georeferenceRemarks
#> 116                 DWC:GEOREFERENCESOURCES                 georeferenceSources
#> 117      DWC:GEOREFERENCEVERIFICATIONSTATUS      georeferenceVerificationStatus
#> 118                               DWC:GROUP                               group
#> 119                             DWC:HABITAT                             habitat
#> 120                DWC:HIGHERCLASSIFICATION                higherClassification
#> 121                     DWC:HIGHERGEOGRAPHY                     higherGeography
#> 122                   DWC:HIGHERGEOGRAPHYID                   higherGeographyID
#> 123         DWC:HIGHESTBIOSTRATIGRAPHICZONE         highestBiostratigraphicZone
#> 124                    DWC:IDENTIFICATIONID                    identificationID
#> 125             DWC:IDENTIFICATIONQUALIFIER             identificationQualifier
#> 126            DWC:IDENTIFICATIONREFERENCES            identificationReferences
#> 127               DWC:IDENTIFICATIONREMARKS               identificationRemarks
#> 128    DWC:IDENTIFICATIONVERIFICATIONSTATUS    identificationVerificationStatus
#> 129                        DWC:IDENTIFIEDBY                        identifiedBy
#> 130                     DWC:INDIVIDUALCOUNT                     individualCount
#> 131                 DWC:INFORMATIONWITHHELD                 informationWithheld
#> 132                DWC:INFRASPECIFICEPITHET                infraspecificEpithet
#> 133                     DWC:INSTITUTIONCODE                     institutionCode
#> 134                       DWC:INSTITUTIONID                       institutionID
#> 135                              DWC:ISLAND                              island
#> 136                         DWC:ISLANDGROUP                         islandGroup
#> 137                             DWC:KINGDOM                             kingdom
#> 138             DWC:LATESTAGEORHIGHESTSTAGE             latestAgeOrHighestStage
#> 139          DWC:LATESTEONORHIGHESTEONOTHEM          latestEonOrHighestEonothem
#> 140          DWC:LATESTEPOCHORHIGHESTSERIES          latestEpochOrHighestSeries
#> 141           DWC:LATESTERAORHIGHESTERATHEM           latestEraOrHighestErathem
#> 142         DWC:LATESTPERIODORHIGHESTSYSTEM         latestPeriodOrHighestSystem
#> 143                           DWC:LIFESTAGE                           lifeStage
#> 144             DWC:LITHOSTRATIGRAPHICTERMS             lithostratigraphicTerms
#> 145                            DWC:LOCALITY                            locality
#> 146                 DWC:LOCATIONACCORDINGTO                 locationAccordingTo
#> 147                          DWC:LOCATIONID                          locationID
#> 148                     DWC:LOCATIONREMARKS                     locationRemarks
#> 149          DWC:LOWESTBIOSTRATIGRAPHICZONE          lowestBiostratigraphicZone
#> 150                    DWC:MATERIALSAMPLEID                    materialSampleID
#> 151                DWC:MAXIMUMDEPTHINMETERS                maximumDepthInMeters
#> 152 DWC:MAXIMUMDISTANCEABOVESURFACEINMETERS maximumDistanceAboveSurfaceInMeters
#> 153            DWC:MAXIMUMELEVATIONINMETERS            maximumElevationInMeters
#> 154                              DWC:MEMBER                              member
#> 155                DWC:MINIMUMDEPTHINMETERS                minimumDepthInMeters
#> 156 DWC:MINIMUMDISTANCEABOVESURFACEINMETERS minimumDistanceAboveSurfaceInMeters
#> 157            DWC:MINIMUMELEVATIONINMETERS            minimumElevationInMeters
#> 158                               DWC:MONTH                               month
#> 159                        DWC:MUNICIPALITY                        municipality
#> 160                     DWC:NAMEACCORDINGTO                     nameAccordingTo
#> 161                   DWC:NAMEACCORDINGTOID                   nameAccordingToID
#> 162                     DWC:NAMEPUBLISHEDIN                     namePublishedIn
#> 163                   DWC:NAMEPUBLISHEDINID                   namePublishedInID
#> 164                 DWC:NAMEPUBLISHEDINYEAR                 namePublishedInYear
#> 165                   DWC:NOMENCLATURALCODE                   nomenclaturalCode
#> 166                 DWC:NOMENCLATURALSTATUS                 nomenclaturalStatus
#> 167                        DWC:OCCURRENCEID                        occurrenceID
#> 168                   DWC:OCCURRENCEREMARKS                   occurrenceRemarks
#> 169                    DWC:OCCURRENCESTATUS                    occurrenceStatus
#> 170                               DWC:ORDER                               order
#> 171                          DWC:ORGANISMID                          organismID
#> 172                        DWC:ORGANISMNAME                        organismName
#> 173                    DWC:ORGANISMQUANTITY                    organismQuantity
#> 174                DWC:ORGANISMQUANTITYTYPE                organismQuantityType
#> 175                     DWC:ORGANISMREMARKS                     organismRemarks
#> 176                       DWC:ORGANISMSCOPE                       organismScope
#> 177                   DWC:ORIGINALNAMEUSAGE                   originalNameUsage
#> 178                 DWC:ORIGINALNAMEUSAGEID                 originalNameUsageID
#> 179                 DWC:OTHERCATALOGNUMBERS                 otherCatalogNumbers
#> 180                DWC:OWNERINSTITUTIONCODE                ownerInstitutionCode
#> 181                       DWC:PARENTEVENTID                       parentEventID
#> 182                     DWC:PARENTNAMEUSAGE                     parentNameUsage
#> 183                   DWC:PARENTNAMEUSAGEID                   parentNameUsageID
#> 184                              DWC:PHYLUM                              phylum
#> 185               DWC:POINTRADIUSSPATIALFIT               pointRadiusSpatialFit
#> 186                        DWC:PREPARATIONS                        preparations
#> 187             DWC:PREVIOUSIDENTIFICATIONS             previousIdentifications
#> 188                          DWC:RECORDEDBY                          recordedBy
#> 189                        DWC:RECORDNUMBER                        recordNumber
#> 190               DWC:REPRODUCTIVECONDITION               reproductiveCondition
#> 191                      DWC:SAMPLESIZEUNIT                      sampleSizeUnit
#> 192                     DWC:SAMPLESIZEVALUE                     sampleSizeValue
#> 193                      DWC:SAMPLINGEFFORT                      samplingEffort
#> 194                    DWC:SAMPLINGPROTOCOL                    samplingProtocol
#> 195                      DWC:SCIENTIFICNAME                      scientificName
#> 196            DWC:SCIENTIFICNAMEAUTHORSHIP            scientificNameAuthorship
#> 197                    DWC:SCIENTIFICNAMEID                    scientificNameID
#> 198                                 DWC:SEX                                 sex
#> 199                     DWC:SPECIFICEPITHET                     specificEpithet
#> 200                      DWC:STARTDAYOFYEAR                      startDayOfYear
#> 201                       DWC:STATEPROVINCE                       stateProvince
#> 202                            DWC:SUBGENUS                            subgenus
#> 203                      DWC:TAXONCONCEPTID                      taxonConceptID
#> 204                             DWC:TAXONID                             taxonID
#> 205                     DWC:TAXONOMICSTATUS                     taxonomicStatus
#> 206                           DWC:TAXONRANK                           taxonRank
#> 207                        DWC:TAXONREMARKS                        taxonRemarks
#> 208                          DWC:TYPESTATUS                          typeStatus
#> 209                 DWC:VERBATIMCOORDINATES                 verbatimCoordinates
#> 210            DWC:VERBATIMCOORDINATESYSTEM            verbatimCoordinateSystem
#> 211                       DWC:VERBATIMDEPTH                       verbatimDepth
#> 212                   DWC:VERBATIMELEVATION                   verbatimElevation
#> 213                   DWC:VERBATIMEVENTDATE                   verbatimEventDate
#> 214                    DWC:VERBATIMLATITUDE                    verbatimLatitude
#> 215                    DWC:VERBATIMLOCALITY                    verbatimLocality
#> 216                   DWC:VERBATIMLONGITUDE                   verbatimLongitude
#> 217                         DWC:VERBATIMSRS                         verbatimSRS
#> 218                   DWC:VERBATIMTAXONRANK                   verbatimTaxonRank
#> 219                      DWC:VERNACULARNAME                      vernacularName
#> 220                           DWC:WATERBODY                           waterBody
#> 221                                DWC:YEAR                                year
#> 222                       DYNAMICPROPERTIES                   dynamicProperties
#> 223                EARLIESTAGEORLOWESTSTAGE            earliestAgeOrLowestStage
#> 224             EARLIESTEONORLOWESTEONOTHEM         earliestEonOrLowestEonothem
#> 225             EARLIESTEPOCHORLOWESTSERIES         earliestEpochOrLowestSeries
#> 226              EARLIESTERAORLOWESTERATHEM          earliestEraOrLowestErathem
#> 227            EARLIESTPERIODORLOWESTSYSTEM        earliestPeriodOrLowestSystem
#> 228                            ENDDAYOFYEAR                        endDayOfYear
#> 229                                 ESPECIE                     specificEpithet
#> 230                      ESTABLISHMENTMEANS                  establishmentMeans
#> 231                                  ESTADO                       stateProvince
#> 232                               EVENTDATE                           eventDate
#> 233                                 EVENTID                             eventID
#> 234                            EVENTREMARKS                        eventRemarks
#> 235                               EVENTTIME                           eventTime
#> 237                                 FAMILIA                              family
#> 238                                  FAMILY                              family
#> 239                        FECHA DE COLECTA                           eventDate
#> 241                              FIELDNOTES                          fieldNotes
#> 242                             FIELDNUMBER                         fieldNumber
#> 243                                    FILO                              phylum
#> 244                     FOOTPRINTSPATIALFIT                 footprintSpatialFit
#> 245                            FOOTPRINTSRS                        footprintSRS
#> 246                            FOOTPRINTWKT                        footprintWKT
#> 247                               FORMATION                           formation
#> 249                                  GENERO                               genus
#> 250                                  GÉNERO                               genus
#> 252                                   GENUS                               genus
#> 253                           GEODETICDATUM                       geodeticDatum
#> 254                     GEOLOGICALCONTEXTID                 geologicalContextID
#> 255                         GEOREFERENCEDBY                     georeferencedBy
#> 256                       GEOREFERENCEDDATE                   georeferencedDate
#> 257                    GEOREFERENCEPROTOCOL                georeferenceProtocol
#> 258                     GEOREFERENCEREMARKS                 georeferenceRemarks
#> 259                     GEOREFERENCESOURCES                 georeferenceSources
#> 260          GEOREFERENCEVERIFICATIONSTATUS      georeferenceVerificationStatus
#> 261                                   GROUP                               group
#> 262                                 HABITAT                             habitat
#> 263                    HIGHERCLASSIFICATION                higherClassification
#> 264                         HIGHERGEOGRAPHY                     higherGeography
#> 265                       HIGHERGEOGRAPHYID                   higherGeographyID
#> 266             HIGHESTBIOSTRATIGRAPHICZONE         highestBiostratigraphicZone
#> 268                        IDENTIFICATIONID                    identificationID
#> 269                 IDENTIFICATIONQUALIFIER             identificationQualifier
#> 270                IDENTIFICATIONREFERENCES            identificationReferences
#> 271                   IDENTIFICATIONREMARKS               identificationRemarks
#> 272        IDENTIFICATIONVERIFICATIONSTATUS    identificationVerificationStatus
#> 273                            IDENTIFIEDBY                        identifiedBy
#> 274                         INDIVIDUALCOUNT                     individualCount
#> 275                     INFORMATIONWITHHELD                 informationWithheld
#> 276                    INFRASPECIFICEPITHET                infraspecificEpithet
#> 277                         INSTITUTIONCODE                     institutionCode
#> 278                           INSTITUTIONID                       institutionID
#> 279                                    ISLA                              island
#> 280                                  ISLAND                              island
#> 281                             ISLANDGROUP                         islandGroup
#> 282                                 KINGDOM                             kingdom
#> 283                                LANGUAGE                            language
#> 284                                     LAT                    verbatimLatitude
#> 285                 LATESTAGEORHIGHESTSTAGE             latestAgeOrHighestStage
#> 286              LATESTEONORHIGHESTEONOTHEM          latestEonOrHighestEonothem
#> 287              LATESTEPOCHORHIGHESTSERIES          latestEpochOrHighestSeries
#> 288               LATESTERAORHIGHESTERATHEM           latestEraOrHighestErathem
#> 289             LATESTPERIODORHIGHESTSYSTEM         latestPeriodOrHighestSystem
#> 290                                 LATITUD                    verbatimLatitude
#> 291                                LATITUDE                    verbatimLatitude
#> 292                                  LENGTH                              length
#> 293                                 LICENSE                             license
#> 294                               LIFESTAGE                           lifeStage
#> 295                 LITHOSTRATIGRAPHICTERMS             lithostratigraphicTerms
#> 296                               LOCALIDAD                            locality
#> 297                                LOCALITY                            locality
#> 298                     LOCATIONACCORDINGTO                 locationAccordingTo
#> 299                              LOCATIONID                          locationID
#> 300                         LOCATIONREMARKS                     locationRemarks
#> 301                                     LON                   verbatimLongitude
#> 302                                    LONG                   verbatimLongitude
#> 303                                LONGITUD                   verbatimLongitude
#> 304                               LONGITUDE                   verbatimLongitude
#> 305                               LOT COUNT                     individualCount
#> 306              LOWESTBIOSTRATIGRAPHICZONE          lowestBiostratigraphicZone
#> 307                        MATERIALSAMPLEID                    materialSampleID
#> 308                    MAXIMUMDEPTHINMETERS                maximumDepthInMeters
#> 309     MAXIMUMDISTANCEABOVESURFACEINMETERS maximumDistanceAboveSurfaceInMeters
#> 310                MAXIMUMELEVATIONINMETERS            maximumElevationInMeters
#> 311                                  MEMBER                              member
#> 312                                     MES                               month
#> 314                    MINIMUMDEPTHINMETERS                minimumDepthInMeters
#> 315     MINIMUMDISTANCEABOVESURFACEINMETERS minimumDistanceAboveSurfaceInMeters
#> 316                MINIMUMELEVATIONINMETERS            minimumElevationInMeters
#> 317                                MODIFIED                            modified
#> 318                                   MONTH                               month
#> 319                          MONTHCOLLECTED                               month
#> 320                         MONTHIDENTIFIED                     monthIdentified
#> 321                            MUNICIPALITY                        municipality
#> 322                               MUNICIPIO                        municipality
#> 323                         NAMEACCORDINGTO                     nameAccordingTo
#> 324                       NAMEACCORDINGTOID                   nameAccordingToID
#> 325                         NAMEPUBLISHEDIN                     namePublishedIn
#> 326                       NAMEPUBLISHEDINID                   namePublishedInID
#> 327                     NAMEPUBLISHEDINYEAR                 namePublishedInYear
#> 328                       NOMBRE CIENTIFICO                      scientificName
#> 329                       NOMBRE CIENTÍFICO                      scientificName
#> 330                       NOMENCLATURALCODE                   nomenclaturalCode
#> 331                     NOMENCLATURALSTATUS                 nomenclaturalStatus
#> 332                                  NUMCAT                       catalogNumber
#> 333                            OCCURRENCEID                        occurrenceID
#> 334                       OCCURRENCEREMARKS                   occurrenceRemarks
#> 335                        OCCURRENCESTATUS                    occurrenceStatus
#> 336                                   OCEAN                           waterBody
#> 337                                  OCEANO                           waterBody
#> 338                                  OCÉANO                           waterBody
#> 339                                   ORDEN                               order
#> 340                                   ORDER                               order
#> 341                              ORGANISMID                          organismID
#> 342                            ORGANISMNAME                        organismName
#> 343                        ORGANISMQUANTITY                    organismQuantity
#> 344                    ORGANISMQUANTITYTYPE                organismQuantityType
#> 345                         ORGANISMREMARKS                     organismRemarks
#> 346                           ORGANISMSCOPE                       organismScope
#> 347                       ORIGINALNAMEUSAGE                   originalNameUsage
#> 348                     ORIGINALNAMEUSAGEID                 originalNameUsageID
#> 349                     OTHERCATALOGNUMBERS                 otherCatalogNumbers
#> 350                    OWNERINSTITUTIONCODE                ownerInstitutionCode
#> 351                                    PAÍS                             country
#> 352                           PARENTEVENTID                       parentEventID
#> 353                         PARENTNAMEUSAGE                     parentNameUsage
#> 354                       PARENTNAMEUSAGEID                   parentNameUsageID
#> 356                                   PHOTO                     associatedMedia
#> 357                                  PHYLUM                              phylum
#> 359                   POINTRADIUSSPATIALFIT               pointRadiusSpatialFit
#> 360                            PREPARATIONS                        preparations
#> 363                 PREVIOUSIDENTIFICATIONS             previousIdentifications
#> 364                        PREVIOUSTISSUEID                 otherCatalogNumbers
#> 366                                PROVINCE                       stateProvince
#> 367                               PROVINCIA                       stateProvince
#> 368                              RECORDEDBY                          recordedBy
#> 369                            RECORDNUMBER                        recordNumber
#> 370                              REFERENCES                          references
#> 371                                   REINO                             kingdom
#> 372                   REPRODUCTIVECONDITION               reproductiveCondition
#> 373                            RIGHTSHOLDER                        rightsHolder
#> 374                                     RIO                           waterBody
#> 375                                     RÍO                           waterBody
#> 376                                   RIVER                           waterBody
#> 377                          SAMPLESIZEUNIT                      sampleSizeUnit
#> 378                         SAMPLESIZEVALUE                     sampleSizeValue
#> 379                          SAMPLINGEFFORT                      samplingEffort
#> 380                        SAMPLINGPROTOCOL                    samplingProtocol
#> 381                          SCIENTIFICNAME                      scientificName
#> 382                SCIENTIFICNAMEAUTHORSHIP            scientificNameAuthorship
#> 383                        SCIENTIFICNAMEID                    scientificNameID
#> 384                                     SEX                                 sex
#> 385                                    SEXO                                 sex
#> 386                                 SPECIES                     specificEpithet
#> 387                         SPECIFICEPITHET                     specificEpithet
#> 388                          STARTDAYOFYEAR                      startDayOfYear
#> 389                                   STATE                       stateProvince
#> 390                          STATE/PROVINCE                       stateProvince
#> 391                           STATEPROVINCE                       stateProvince
#> 392                              SUBESPECIE                infraspecificEpithet
#> 394                                SUBGENUS                            subgenus
#> 395                              SUBSPECIES                infraspecificEpithet
#> 397                          TAXONCONCEPTID                      taxonConceptID
#> 398                                 TAXONID                             taxonID
#> 399                         TAXONOMICSTATUS                     taxonomicStatus
#> 400                               TAXONRANK                           taxonRank
#> 401                            TAXONREMARKS                        taxonRemarks
#> 403                                    TYPE                                type
#> 404                              TYPESTATUS                          typeStatus
#> 405                     VERBATIMCOORDINATES                 verbatimCoordinates
#> 406                VERBATIMCOORDINATESYSTEM            verbatimCoordinateSystem
#> 407                           VERBATIMDEPTH                       verbatimDepth
#> 408                       VERBATIMELEVATION                   verbatimElevation
#> 409                       VERBATIMEVENTDATE                   verbatimEventDate
#> 410                        VERBATIMLATITUDE                    verbatimLatitude
#> 411                        VERBATIMLOCALITY                    verbatimLocality
#> 412                       VERBATIMLONGITUDE                   verbatimLongitude
#> 413                             VERBATIMSRS                         verbatimSRS
#> 414                       VERBATIMTAXONRANK                   verbatimTaxonRank
#> 415                          VERNACULARNAME                      vernacularName
#> 416                               WATERBODY                           waterBody
#> 417                                  WEIGHT                              weight
#> 419                                 WORMSID                             taxonID
#> 420                                    YEAR                                year
#> 421                           YEARCOLLECTED                                year
#> 422                          YEARIDENTIFIED                      yearIdentified